+++ src/drawing.c
@@ -225,7 +225,7 @@ _drawing_write_to(lxw_drawing *self, lxw_drawing_coords *coords)
  * Write the <xdr:cNvPr> element.
  */
 STATIC void
-_drawing_write_c_nv_pr(lxw_drawing *self, char *object_name, uint16_t index,
+_drawing_write_c_nv_pr(lxw_drawing *self, char *object_name, uint32_t index,
                        lxw_drawing_object *drawing_object)
 {
     struct xml_attribute_list attributes;
@@ -282,7 +282,7 @@ _drawing_write_c_nv_pic_pr(lxw_drawing *self)
  * Write the <xdr:nvPicPr> element.
  */
 STATIC void
-_drawing_write_nv_pic_pr(lxw_drawing *self, uint16_t index,
+_drawing_write_nv_pic_pr(lxw_drawing *self, uint32_t index,
                          lxw_drawing_object *drawing_object)
 {
     lxw_xml_start_tag(self->file, "xdr:nvPicPr", NULL);
@@ -300,7 +300,7 @@ _drawing_write_nv_pic_pr(lxw_drawing *self, uint16_t index,
  * Write the <a:blip> element.
  */
 STATIC void
-_drawing_write_a_blip(lxw_drawing *self, uint16_t index)
+_drawing_write_a_blip(lxw_drawing *self, uint32_t index)
 {
     struct xml_attribute_list attributes;
     struct xml_attribute *attribute;
@@ -345,7 +345,7 @@ _drawing_write_a_stretch(lxw_drawing *self)
  * Write the <xdr:blipFill> element.
  */
 STATIC void
-_drawing_write_blip_fill(lxw_drawing *self, uint16_t index)
+_drawing_write_blip_fill(lxw_drawing *self, uint32_t index)
 {
     lxw_xml_start_tag(self->file, "xdr:blipFill", NULL);
 
@@ -463,7 +463,7 @@ _drawing_write_sp_pr(lxw_drawing *self, lxw_drawing_object *drawing_object)
  * Write the <xdr:pic> element.
  */
 STATIC void
-_drawing_write_pic(lxw_drawing *self, uint16_t index,
+_drawing_write_pic(lxw_drawing *self, uint32_t index,
                    lxw_drawing_object *drawing_object)
 {
     lxw_xml_start_tag(self->file, "xdr:pic", NULL);
@@ -529,7 +529,7 @@ _drawing_write_c_nv_graphic_frame_pr(lxw_drawing *self)
  * Write the <xdr:nvGraphicFramePr> element.
  */
 STATIC void
-_drawing_write_nv_graphic_frame_pr(lxw_drawing *self, uint16_t index)
+_drawing_write_nv_graphic_frame_pr(lxw_drawing *self, uint32_t index)
 {
     lxw_xml_start_tag(self->file, "xdr:nvGraphicFramePr", NULL);
 
@@ -599,7 +599,7 @@ _drawing_write_xfrm(lxw_drawing *self)
  * Write the <c:chart> element.
  */
 STATIC void
-_drawing_write_chart(lxw_drawing *self, uint16_t index)
+_drawing_write_chart(lxw_drawing *self, uint32_t index)
 {
     struct xml_attribute_list attributes;
     struct xml_attribute *attribute;
@@ -623,7 +623,7 @@ _drawing_write_chart(lxw_drawing *self, uint16_t index)
  * Write the <a:graphicData> element.
  */
 STATIC void
-_drawing_write_a_graphic_data(lxw_drawing *self, uint16_t index)
+_drawing_write_a_graphic_data(lxw_drawing *self, uint32_t index)
 {
     struct xml_attribute_list attributes;
     struct xml_attribute *attribute;
@@ -646,7 +646,7 @@ _drawing_write_a_graphic_data(lxw_drawing *self, uint16_t index)
  * Write the <a:graphic> element.
  */
 STATIC void
-_drawing_write_a_graphic(lxw_drawing *self, uint16_t index)
+_drawing_write_a_graphic(lxw_drawing *self, uint32_t index)
 {
 
     lxw_xml_start_tag(self->file, "a:graphic", NULL);
@@ -661,7 +661,7 @@ _drawing_write_a_graphic(lxw_drawing *self, uint16_t index)
  * Write the <xdr:graphicFrame> element.
  */
 STATIC void
-_drawing_write_graphic_frame(lxw_drawing *self, uint16_t index)
+_drawing_write_graphic_frame(lxw_drawing *self, uint32_t index)
 {
     struct xml_attribute_list attributes;
     struct xml_attribute *attribute;
@@ -689,7 +689,7 @@ _drawing_write_graphic_frame(lxw_drawing *self, uint16_t index)
  * Write the <xdr:twoCellAnchor> element.
  */
 STATIC void
-_drawing_write_two_cell_anchor(lxw_drawing *self, uint16_t index,
+_drawing_write_two_cell_anchor(lxw_drawing *self, uint32_t index,
                                lxw_drawing_object *drawing_object)
 {
     struct xml_attribute_list attributes;
@@ -811,7 +811,7 @@ _drawing_write_absolute_anchor(lxw_drawing *self)
 void
 lxw_drawing_assemble_xml_file(lxw_drawing *self)
 {
-    uint16_t index;
+    uint32_t index;
     lxw_drawing_object *drawing_object;
 
     /* Write the XML declaration. */
